Add Columns constrains

+++ src/main/java/com/abutua/productbackend/resources/ProductController.java
@@ -41,7 +41,7 @@ public class ProductController {
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<Product> getProduct(@PathVariable int id) {
+    public ResponseEntity<Product> getProduct(@PathVariable long id) {
         Product product = productService.getById(id);
         return ResponseEntity.ok(product);
     }
@@ -52,13 +52,13 @@ public class ProductController {
     }
 
     @DeleteMapping("{id}")    
-    public ResponseEntity<Void> removeProduct(@PathVariable int id) {
+    public ResponseEntity<Void> removeProduct(@PathVariable long id) {
         productService.deleteById(id);                
         return ResponseEntity.noContent().build(); 
     }
    
     @PutMapping("{id}")    
-    public ResponseEntity<Void> updateProduct(@PathVariable int id, @RequestBody Product productUpdate) {
+    public ResponseEntity<Void> updateProduct(@PathVariable long id, @RequestBody Product productUpdate) {
         productService.update(id, productUpdate);
         return ResponseEntity.ok().build(); 
     }
